Move new rf-specific app to rf/app
