+++ sigpy/mri/__init__.py
@@ -9,7 +9,6 @@ and density compensation factors.
 from sigpy.mri import app, linop
 
 from sigpy.mri import bloch, dcf, precond, samp, sim, util
-from sigpy.mri import bloch, precond, rf, samp, sim, util
 from sigpy.mri.bloch import *  # noqa
 from sigpy.mri.dcf import *  # noqa
 from sigpy.mri.precond import *  # noqa
