+++ sigpy/mri/rf/__init__.py
@@ -6,9 +6,8 @@ provides functions to compute preconditioners.
 
 """
 from sigpy.mri import app, linop
-from sigpy.mri.rf import app
 
-from sigpy.mri.rf import ptx, slr, optcont, adiabatic, sim, b1sel
+from sigpy.mri.rf import ptx, slr, optcont, adiabatic, sim, b1sel, trajgrad
 from sigpy.mri.rf.ptx import *  # noqa
 from sigpy.mri.rf.slr import *  # noqa
 from sigpy.mri.rf.optcont import *  # noqa
